refactor(canvas): establish main model domain
