+++ apps/canvas-workspace/src/main/__tests__/import-boundaries.test.ts
@@ -147,6 +147,7 @@ describe('import boundaries', () => {
 const FORBIDDEN_MAIN_DOMAIN_IMPORTS = new Map<string, Set<string>>([
   ['agent', new Set(['app', 'scheduled'])],
   ['canvas', new Set(['agent'])],
+  ['generation', new Set(['agent'])],
   ['plugin-market', new Set(['agent'])],
   ['settings', new Set(['plugin-market'])],
   ['webview', new Set(['agent'])],
