+++ apps/canvas-workspace/src/main/agent/prepared-chat.ts
@@ -3,7 +3,7 @@ import type { WebContents } from 'electron';
 
 import type { AgentRequestContext, AgentScope } from './types';
 import type { CanvasAgentService } from './service';
-import type { ResolvedCanvasModel } from './model/config';
+import type { ResolvedCanvasModel } from '../models/config';
 import { isPerfChatReplayRequest, replayPerfChatStream } from './perf-chat-replay';
 import { GLOBAL_CHAT_SESSION_STORE_ID, SessionStore } from './session-store';
 import { ActiveChatRegistry, type ChatRunStreamChannel } from './active-chat-registry';
